+++ elide-core/src/main/java/com/yahoo/elide/core/EntityDictionary.java
@@ -34,8 +34,6 @@ import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
-import java.lang.reflect.ParameterizedType;
-import java.lang.reflect.Type;
 import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -44,6 +42,7 @@ import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.Queue;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
@@ -235,7 +234,7 @@ public class EntityDictionary {
 
         if (checkCls == null) {
             try {
-                checkCls = (Class<? extends Check>) Class.forName(checkIdentifier);
+                checkCls = Class.forName(checkIdentifier).asSubclass(Check.class);
                 try {
                     checkNames.putIfAbsent(checkIdentifier, checkCls);
                 } catch (IllegalArgumentException e) {
@@ -631,19 +630,7 @@ public class EntityDictionary {
             return null;
         }
 
-        Type type;
-
-        if (fieldOrMethod instanceof Method) {
-            type = ((Method) fieldOrMethod).getGenericReturnType();
-        } else {
-            type = ((Field) fieldOrMethod).getGenericType();
-        }
-
-        if (type instanceof ParameterizedType) {
-            return (Class<?>) ((ParameterizedType) type).getActualTypeArguments()[paramIndex];
-        }
-
-        return getType(entityClass, identifier);
+        return EntityBinding.getFieldType(entityClass, fieldOrMethod, Optional.of(paramIndex));
     }
 
     /**
@@ -1127,7 +1114,7 @@ public class EntityDictionary {
     /**
      * Returns whether or not a class is already bound.
      * @param cls
-     * @return
+     * @return true if the class is bound.  False otherwise.
      */
     public boolean hasBinding(Class<?> cls) {
         return bindJsonApiToEntity.contains(cls);