+++ elide-core/src/test/java/com/yahoo/elide/core/EntityDictionaryTest.java
@@ -14,6 +14,8 @@ import com.yahoo.elide.annotation.MappedInterface;
 import com.yahoo.elide.annotation.OnUpdatePreSecurity;
 import com.yahoo.elide.annotation.ReadPermission;
 import com.yahoo.elide.functions.LifeCycleHook;
+import com.yahoo.elide.models.generics.Employee;
+import com.yahoo.elide.models.generics.Manager;
 import com.yahoo.elide.security.checks.prefab.Collections.AppendOnly;
 import com.yahoo.elide.security.checks.prefab.Collections.RemoveOnly;
 import com.yahoo.elide.security.checks.prefab.Common.UpdateOnCreate;
@@ -67,6 +69,8 @@ public class EntityDictionaryTest extends EntityDictionary {
         this.bindEntity(StringId.class);
         this.bindEntity(Friend.class);
         this.bindEntity(FieldAnnotations.class);
+        this.bindEntity(Manager.class);
+        this.bindEntity(Employee.class);
 
         checkNames.forcePut("user has all access", Role.ALL.class);
     }
@@ -252,6 +256,15 @@ public class EntityDictionaryTest extends EntityDictionary {
 
         type = getParameterizedType(fun, "relation3");
         Assert.assertEquals(type, Child.class, "A Child object should return Child.class");
+
+        Assert.assertEquals(getParameterizedType(FieldAnnotations.class, "children"), FieldAnnotations.class,
+                "getParameterizedType return the type of a private field relationship");
+
+        Assert.assertEquals(getParameterizedType(Parent.class, "children"), Child.class,
+            "getParameterizedType returns the type of relationship fields");
+
+        Assert.assertEquals(getParameterizedType(Manager.class, "minions"), Employee.class,
+            "getParameterizedType returns the correct generic type of a to-many relationship");
     }
 
     @Test
@@ -354,7 +367,6 @@ public class EntityDictionaryTest extends EntityDictionary {
 
     @Test
     public void testGetType() throws Exception {
-
         Assert.assertEquals(getType(FieldAnnotations.class, "id"), Long.class,
             "getType returns the type of the ID field of the given class");
 
@@ -379,6 +391,12 @@ public class EntityDictionaryTest extends EntityDictionary {
 
         Assert.assertEquals(getType(Friend.class, "name"), String.class,
                 "getType returns the type of attribute when defined in a super class");
+
+        Assert.assertEquals(getType(Employee.class, "boss"), Manager.class,
+            "getType returns the correct generic type of a to-one relationship");
+
+        Assert.assertEquals(getType(Manager.class, "minions"), Set.class,
+            "getType returns the correct generic type of a to-many relationship");
     }
 
     @Test(expectedExceptions = IllegalArgumentException.class)