+++ elide-datastore/elide-datastore-hibernate3/src/test/java/com/yahoo/elide/datastores/hibernate3/HibernateDataStoreSupplier.java
@@ -6,6 +6,7 @@
 package com.yahoo.elide.datastores.hibernate3;
 
 import com.yahoo.elide.core.DataStore;
+import com.yahoo.elide.models.generics.Manager;
 import com.yahoo.elide.utils.ClassScanner;
 
 import example.Filtered;
@@ -38,6 +39,8 @@ public class HibernateDataStoreSupplier implements Supplier<DataStore> {
         Configuration configuration = new Configuration();
         try {
             ClassScanner.getAnnotatedClasses(Parent.class.getPackage(), Entity.class)
+                    .forEach(configuration::addAnnotatedClass);
+            ClassScanner.getAnnotatedClasses(Manager.class.getPackage(), Entity.class)
                     .forEach(configuration::addAnnotatedClass);
         } catch (MappingException e) {
             throw new RuntimeException(e);