+++ elide-example-models/src/main/java/com/yahoo/elide/models/generics/package-info.java
@@ -1,0 +1,10 @@
+/*
+ * Copyright 2019, Yahoo Inc.
+ * Licensed under the Apache License, Version 2.0
+ * See LICENSE file in project root for terms.
+ */
+
+@SharePermission
+package com.yahoo.elide.models.generics;
+
+import com.yahoo.elide.annotation.SharePermission;