+++ elide-integration-tests/src/test/java/com/yahoo/elide/initialization/AbstractIntegrationTestInitializer.java
@@ -58,10 +58,11 @@ public abstract class AbstractIntegrationTestInitializer extends AbstractApiReso
     public static DataStore getNewDatabaseManager() {
         try {
             final String dataStoreSupplierName = System.getProperty("dataStoreSupplier");
+            @SuppressWarnings("unchecked")
             Supplier<DataStore> dataStoreSupplier =
-                    (Supplier<DataStore>) Class.forName(dataStoreSupplierName).newInstance();
+                    Class.forName(dataStoreSupplierName).asSubclass(Supplier.class).newInstance();
             dataStore = dataStoreSupplier.get();
-        } catch (InstantiationException | IllegalAccessException | ClassNotFoundException e) {
+        } catch (InstantiationException | IllegalAccessException | ClassNotFoundException | ClassCastException e) {
             throw new IllegalStateException(e);
         }
         return dataStore;